Commit Final : Personne, Enseignant

+++ tp-entreprise-jpa/src/main/java/sopra/entreprise/repository/IPersonneRepository.java
@@ -2,6 +2,6 @@ package sopra.entreprise.repository;
 
 import sopra.entreprise.model.Personne;
 
-public interface IPersonneRepository extends IRepository<Personne, Long> {
+public interface IPersonneRepository extends IRepository<Personne, Integer> {
 
 }
